Expand control of speech engine with Frequency or Morph CV are patched
In the original Plaits firmware, two special controls for the Speech engine (prosody and speed) are exposed via the Frequency and Morph CV atteunverters/trimpots only if the Trigger input is patched AND the Frequency or Morph CV are unpatched, respectively.

This is because these parameters are different from those controlled by the Frequency and Morph knobs (pitch and phoneme/word selection), and thus different from those which should be controlled by CV at the Frequency and Morph inputs.  So in a situation where those inputs are patched the knobs should revert to the expected case of attenuverting control over pitch and phoneme/word selection.

They are also only relevant in situations where the engine is being triggered rather than droning, although the LPG is not actually applied to this engine the controls are exposed in the same situations as the LPG modulation controls.

Given that the Poly_AudibleInstruments version of Macro Oscillator 2 provides extra trimpots, this patch allows for control of Speech engine prosody and speed via the LPG->Frequency/Morph trimpots when Trigger input is patched even when Frequency and/or Morph inputs are also patched.